Alphabet refactoring
Introduced a new alphabet representation based on ints.
This led to a massive refactoring, consisting in rewrite per-functor
input signatures instead of defining all functors around the same set
of core signatures. This has the benefit of making modules more
indenpendent and also it's clearer what a functor really needs.