Linear_algebra.{vector, matrix} are now private
This has been tremendously useful to catch dimension bugs. Vectors are
here defined as matrices of size (n, 1), and the code was full of
places where vectors of shape [| n |] (row vectors) instead of
[| n ; 1 |].

Maybe the private qualifier can be removed in the future, but for now
it's "grave" useful!