Phylo_ctmc: fixed source of infinite loop in substitution mapping sampler
I spotted an infinite loop in some instances, coming from
`conditional_simulation_along_branch_by_uniformization`. The algorithm
in this function samples a number of event initial state `a`, last
state `b`, and the branch length `lambda`. It does so by computing the
marginal probability P[b | a, lambda], sampling some `g` betweenn 0
and 1, and computing P[b, n | a, lambda] for increasing `n` until
reaching the `g` * P[b | a, lambda].

However, in the case P[b | a, lambda] is very small, the values for
P[b, n | a, lambda] can lack precision and they don't sum close enough
to P[b | a, lambda]. So we end up computing P[b, n | a, lambda] for
increasing n forever.

The proposed fix is to required a `max_path_length` argument that
gives an upper bound on the number of events on the branch. When
reaching this, we sample the number of events on the branch in a
different, slightly less efficient way, that assume that the number of
events on the branch is upper bounded.